Avoid overiding values extracted form different headers
If the same mailto is extracted from diffent parts of the header, the settings from the last extraction overrides any other extraction.

Example say you have
```$rcmail_config['custom_from_header_rules'] = 'Envelope-To=deo;to=de';```
And the header contains
```To: me@bar.com ... Envelope-To: me@bar.com```
And your domain is
```foo.com```

The address me@bar.com will not be choosen because the settings for me@bar.com will both be saved in ```$recipients['me@bar.com']``` with the one for the To-Header overriding the one for the Envelope-To-Header. That results in not taking me@bar.com becaues ```$recipients['me@bar.com']['match_other']``` will be false.