Issue #18 - Adjust Classic skin's compose-div when toggling custom from input
In the Classic skin, the custom from input box takes up the whole width of the
layout, so the additional control for toggling the custom from and the settings
link end up on a new line.

Because the Classic skin uses rather a lot of absolute positioning, this means
that we need to move the editor textarea out of the way ourselves, so the
subject line doesn't get pushed underneath it.